# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Geography Quiz platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/progress:
    get:
      operationId: getUserProgress
      tags: [user-progress]
      summary: Get the current user's quiz attempt history and stats
      responses:
        "200":
          description: User progress summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProgress"
        "401":
          description: Unauthorized
...
  /user/progress/{quizId}:
    get:
      operationId: getUserQuizProgress
      tags: [user-progress]
      summary: Get the current user's history for a specific quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User quiz history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserQuizProgress"
        "401":
          description: Unauthorized
components:
  schemas:
...
    UserProgress:
      type: object
      properties:
        totalAttempts:
          type: integer
        totalQuizzesTaken:
          type: integer
        averagePercentage:
          type: number
        bestPercentage:
          type: number
        recentAttempts:
          type: array
          items:
            $ref: "#/components/schemas/UserAttemptSummary"
      required:
        - totalAttempts
        - totalQuizzesTaken
        - averagePercentage
        - bestPercentage
        - recentAttempts
    UserAttemptSummary:
      type: object
      properties:
        id:
          type: integer
        quizId:
          type: integer
        quizTitle:
          type: string
        score:
          type: integer
        totalQuestions:
          type: integer
        percentage:
          type: number
        completedAt:
          type: string
      required:
        - id
        - quizId
        - quizTitle
        - score
        - totalQuestions
        - percentage
        - completedAt
    UserQuizProgress:
      type: object
      properties:
        quizId:
          type: integer
        attempts:
          type: integer
        bestScore:
          type: integer
        bestPercentage:
          type: number
        lastAttemptAt:
          type: ["string", "null"]
        history:
          type: array
          items:
            $ref: "#/components/schemas/UserAttemptSummary"
      required:
        - quizId
        - attempts
        - bestScore
        - bestPercentage
        - lastAttemptAt
        - history
...